requirements.txt
playwright
pandas
streamlit
supabase
python-dotenv
openpyxl
plotly
